requirements.txt
Django
django-probes
djangorestframework
django-rest-swagger
django-cors-headers
django-filter
drf_spectacular
django-iris